Addition of textured strips, music tweaking
A big addition to the visuals of the game this time: textured strips, which are patches on the ground that are, of course, textured. This can therefore give more depth to the game world. Also tweaked the music a bit (yet again), and attempted to change the way music samples are queued, but that part is not yet finished.