Adding new OPNA emulation core, debug tool
Implemented a new OPNA emulation core to ensure licence compatibility; there are still some things to fix before I remove the old core. Added a register viewer for the OPNA emulation for debugging purposes. And there are some miscellaneous fixes as well.
New emulation core key-repeat behaviour still has to be fixed: there are occasional key-on misses, but not enough to be totally unacceptable.