Code pruning and new config options
Removing license-incompatible code and useless code in the Bamboo Tracker Player plugin, and reconfiguring code to work without the missing code. This means that the player no longer can choose the underlying chip emulator, but can now toggle the emulation on/off if they wish to save processing power. Also added accelerator inversion, where one can choose if they want the accelerator to default to on and for the accelerator button to release the accelerator.