Finished new collision logic
I think I've finished the new collision logic for now. Hitting into a static obstacle will now at high enough speed fling you off and onto the ground, which is very humiliating. Added the basics of a system to change the view of rivals depending on their position, but the graphics are not ready yet.